+++ compiler/src/main/java/info/teksol/mc/mindcode/compiler/optimization/ArrayOptimizer.java
@@ -5,6 +5,7 @@ import info.teksol.mc.mindcode.compiler.generation.variables.ArrayStore;
 import info.teksol.mc.mindcode.logic.instructions.ArrayAccessInstruction;
 import info.teksol.mc.mindcode.logic.instructions.ArrayOrganization;
 import info.teksol.mc.mindcode.logic.instructions.LogicInstruction;
+import info.teksol.mc.mindcode.logic.opcodes.Opcode;
 import org.jetbrains.annotations.NotNull;
 import org.jspecify.annotations.NullMarked;
 import org.jspecify.annotations.Nullable;
@@ -51,11 +52,12 @@ class ArrayOptimizer extends BaseOptimizer {
         return false;
     }
 
-    private static @NotNull ArrayOrganization computeArrayOrganization(ArrayAccessInstruction ix) {
+    private @NotNull ArrayOrganization computeArrayOrganization(ArrayAccessInstruction ix) {
         ArrayOrganization current = ix.getArrayOrganization();
         return switch (ix.getArray().getArrayStore().getSize()) {
             case 1 -> ArrayOrganization.INTERNAL_SIZE1;
             case 2 -> ArrayOrganization.INTERNAL_SIZE2;
+            case 3 -> experimental() && instructionProcessor.isSupported(Opcode.SELECT) ? ArrayOrganization.INTERNAL_SIZE3 : current;
             default -> current;
         };
     }