+++ compiler/src/main/java/info/teksol/mc/mindcode/compiler/optimization/OptimizationCoordinator.java
@@ -31,6 +31,7 @@ import static info.teksol.mc.mindcode.compiler.optimization.OptimizationPhase.*;
 public class OptimizationCoordinator {
     public static final boolean TRACE = false;
     public static final boolean DEBUG_PRINT = TRACE;
+    public static final boolean TRACE_ALL = true;
     public static final boolean SYSTEM_OUT = false;
 
     public static final boolean IGNORE_UNINITIALIZED = false;